Actually fix pushing custom tags to ECR :-/ (#39)
The syntax for tag in this case means tag the local image with the specified tag to be uploaded to a particular location. As such we need to tag the local image with the commit tag (as set in `docker build --tag ${CIRCLE_SHA1}` in the build set) but tell it that it's going to be pushed to the `latest` tag in the remote repo.